python/pyproject.toml
[project]
name = "inference-rs-py"
version = "0.1.0"
description = "OpenVINO inference toolkit — Python reference implementation for benchmarking against the Rust version"
requires-python = ">=3.11"
dependencies = [
    "openvino>=2026.0",
    "numpy>=1.26",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[project.scripts]
inference-rs-py = "inference_rs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["inference_rs"]
